Make total in OrderItem float so taxes are rounded after they are summed for all items

+++ src/Sylius/Component/Order/Model/OrderItemInterface.php
@@ -28,7 +28,7 @@ interface OrderItemInterface extends AdjustableInterface, OrderAwareInterface, R
 
     public function setOriginalUnitPrice(?float $originalUnitPrice): void;
 
-    public function getTotal(): int;
+    public function getTotal(): float;
 
     /**
      * Recalculate totals. Should be used after every unit change.
